Renamed files for `JoltCustomUserDataShape`
